Added two stimuli: Image Target: moves an image in 2D in the same way as RectTarget does. Paloma ... : replicating the receptive field determination stimulus used in Paloma's lab.
Starfield, Target 3D:
Changed handling of dot size, so that dots (target or starfield) with the same dot size should be comparable across setups (i.e. the same visual experience for the fly).

For Starfield, added a "debug" option to show a fixed grid of dots - occasionally useful for checking that changes are correct.

Also split Target 3D into two:
The "pursuit" target works as Target 3D has always worked: the distance is direct linear distance from the fly's eye to the target.
The "plain" one (called just Target 3D) treats distance as the distance from the eye to the point on the XY plane that it is looking at. We assume that the fly is looking perpendicularly to the XY plane, and straight along the negative Z axis. Hence distance in this case is the negative of the Z coordinate of the target.

Also added the beginnings of Background2D as a stimulus, but not completed yet, so do not use. (And may be discarded later).